Dropped unused offsetX & offsetY of Asteroid

+++ oni-seed-browser-model/src/commonMain/kotlin/de/stefan_oltmann/oni/model/server/upload/UploadStarMapEntryVanilla.kt
@@ -27,5 +27,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class UploadStarMapEntryVanilla(
     val id: VanillaSpacePOI,
-    val distance: Int
+    val distance: Byte
 )
